ygopro: translate and fix core duel logic
- Fix duel_palyer.go CTOS_CREATE_GAME (packet len check, TAG mode, ETimer)
- Create timer.go (migrate timerWheel from single_duel.go)
- Create replay.go (full Replay struct with record/read/write/LZMA)
- Complete tag_duel.go (~1900 lines, all TagDuel methods)
- Fix single_duel.go (remove duplicate timer, clean up)
- Update protocol types and network layer
- Adjust go.mod/go.sum for new dependencies

+++ ocgcore/api_cgo.go
@@ -36,7 +36,7 @@ func goScriptReader(scriptName *C.char, slen *C.int) *C.uchar {
 }
 
 //export goMessageHandler
-func goMessageHandler(data C.longlong, size C.uint32_t) {
+func goMessageHandler(data C.intptr_t, size C.uint32_t) {
 	_messageHandler(data, size)
 }
 
@@ -68,7 +68,7 @@ func goCardReader(cardID C.uint32_t, data *C.card_data) C.uint32_t {
 type (
 	ScriptReader   func(scriptNamePtr *C.char, slen *C.int) *C.uchar
 	CardReader     func(cardId uint32) *CardData
-	MessageHandler func(data C.longlong, size C.uint32_t)
+	MessageHandler func(data C.intptr_t, size C.uint32_t)
 )
 type Option func(api *OCGApi)
 
@@ -151,7 +151,7 @@ func (o *OCGApi) defaultCardReader(code uint32) *CardData {
 
 	return nil
 }
-func (o *OCGApi) defaultOnMessageHandler(duelPtr C.longlong, msgType C.uint32_t) {
+func (o *OCGApi) defaultOnMessageHandler(duelPtr C.intptr_t, msgType C.uint32_t) {
 	duelLock.Lock()
 	defer duelLock.Unlock()
 	duel, has := duels[uintptr(duelPtr)]
@@ -164,47 +164,47 @@ func (o *OCGApi) CreateDuel(seed int32) uintptr {
 	return uintptr(C.create_duel(C.int32_t(seed)))
 }
 func (o *OCGApi) StartDuel(pduel uintptr, options int32) {
-	C.start_duel(C.longlong(pduel), C.int32_t(options))
+	C.start_duel(C.intptr_t(pduel), C.int32_t(options))
 }
 func (o *OCGApi) EndDuel(pduel uintptr) {
-	C.end_duel(C.longlong(pduel))
+	C.end_duel(C.intptr_t(pduel))
 }
 func (o *OCGApi) SetPlayerInfo(pduel uintptr, playerId, LP, startCount, drawCount int32) {
-	C.set_player_info(C.longlong(pduel), C.int32_t(playerId), C.int32_t(LP), C.int32_t(startCount), C.int32_t(drawCount))
+	C.set_player_info(C.intptr_t(pduel), C.int32_t(playerId), C.int32_t(LP), C.int32_t(startCount), C.int32_t(drawCount))
 }
 func (o *OCGApi) GetLogMessage(pduel uintptr, buf []byte) {
-	C.get_log_message(C.longlong(pduel), (*C.uchar)(unsafe.Pointer(&buf[0])))
+	C.get_log_message(C.intptr_t(pduel), (*C.uchar)(unsafe.Pointer(&buf[0])))
 }
 func (o *OCGApi) GetMessage(pduel uintptr, buff []byte) int32 {
-	return int32(C.get_message(C.longlong(pduel), (*C.uchar)(unsafe.Pointer(&buff[0]))))
+	return int32(C.get_message(C.intptr_t(pduel), (*C.uchar)(unsafe.Pointer(&buff[0]))))
 }
 func (o *OCGApi) Process(pduel uintptr) uint32 {
-	return uint32(C.process(C.longlong(pduel)))
+	return uint32(C.process(C.intptr_t(pduel)))
 }
 func (o *OCGApi) NewCard(pduel uintptr, code uint32, owner, playerid, location, sequence, position uint8) {
-	C.new_card(C.longlong(pduel), C.uint32_t(code), C.uint8_t(owner), C.uint8_t(playerid), C.uint8_t(location), C.uint8_t(sequence), C.uint8_t(position))
+	C.new_card(C.intptr_t(pduel), C.uint32_t(code), C.uint8_t(owner), C.uint8_t(playerid), C.uint8_t(location), C.uint8_t(sequence), C.uint8_t(position))
 }
 func (o *OCGApi) NewTagCard(pduel uintptr, code uint32, owner, location uint8) {
 
 }
 func (o *OCGApi) QueryCard(pduel uintptr, playerid, location, sequence uint8, queryFlag int32, buf []byte, useCache int32) int32 {
-	return int32(C.query_card(C.longlong(pduel), C.uint8_t(playerid), C.uint8_t(location), C.uint8_t(sequence), C.int32_t(queryFlag), (*C.uchar)(unsafe.Pointer(&buf[0])), C.int32_t(useCache)))
+	return int32(C.query_card(C.intptr_t(pduel), C.uint8_t(playerid), C.uint8_t(location), C.uint8_t(sequence), C.int32_t(queryFlag), (*C.uchar)(unsafe.Pointer(&buf[0])), C.int32_t(useCache)))
 }
 func (o *OCGApi) QueryFieldCount(pduel uintptr, playerid, location uint8) int32 {
-	return int32(C.query_field_count(C.longlong(pduel), C.uint8_t(playerid), C.uint8_t(location)))
+	return int32(C.query_field_count(C.intptr_t(pduel), C.uint8_t(playerid), C.uint8_t(location)))
 }
 func (o *OCGApi) QueryFieldCard(pduel uintptr, playerId, location uint8, queryFlag uint32, buf []byte, useCache int32) int32 {
-	return int32(C.query_field_card(C.longlong(pduel), C.uint8_t(playerId), C.uint8_t(location), C.int32_t(queryFlag), (*C.uchar)(unsafe.Pointer(&buf[0])), C.int32_t(useCache)))
+	return int32(C.query_field_card(C.intptr_t(pduel), C.uint8_t(playerId), C.uint8_t(location), C.int32_t(queryFlag), (*C.uchar)(unsafe.Pointer(&buf[0])), C.int32_t(useCache)))
 }
 func (o *OCGApi) QueryFieldInfo(pduel uintptr, buf []byte) int32 {
-	return int32(C.query_field_info(C.longlong(pduel), (*C.uchar)(unsafe.Pointer(&buf[0]))))
+	return int32(C.query_field_info(C.intptr_t(pduel), (*C.uchar)(unsafe.Pointer(&buf[0]))))
 }
 func (o *OCGApi) SetResponseI(pduel uintptr, value int32) {
-	C.set_responsei(C.longlong(pduel), C.int32_t(value))
+	C.set_responsei(C.intptr_t(pduel), C.int32_t(value))
 }
 func (o *OCGApi) SetResponseB(pduel uintptr, buf []byte) {
-	C.set_responseb(C.longlong(pduel), (*C.uchar)(unsafe.Pointer(&buf[0])))
+	C.set_responseb(C.intptr_t(pduel), (*C.uchar)(unsafe.Pointer(&buf[0])))
 }
 func (o *OCGApi) PreloadScript(pduel uintptr, script []byte) int32 {
-	return int32(C.preload_script(C.longlong(pduel), (*C.char)(unsafe.Pointer(&script[0])), C.int32_t(len(script))))
+	return int32(C.preload_script(C.intptr_t(pduel), (*C.char)(unsafe.Pointer(&script[0])), C.int32_t(len(script))))
 }
